% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_diccionario.R
\name{edit_dicc}
\alias{edit_dicc}
\title{Diccionarios EDIT-S}
\usage{
edit_dicc(periodo = NULL, panel = NULL, col_types = "cccc")
}
\arguments{
\item{periodo}{una cadena de texto. Selecciona el periodo del diccionario.}

\item{panel}{una cadena de texto. Selecciona el panel del diccionario.}

\item{col_types}{una cadena de texto. Especifica el tipo de columnas que compone la base de datos del diccionario.}
}
\value{
un data frame con el diccionario seleccionado.
}
\description{
Carga un diccionario de la base de datos EDIT-S al ambiente global
}
\examples{
edit_diccionario(periodo = "ii")

}
