% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_panel.R
\name{edit_panel}
\alias{edit_panel}
\title{Construye paneles a partir de los datos EDIT-S}
\usage{
edit_panel(panel, capitulo)
}
\arguments{
\item{panel}{una cadena de texto. Selecciona el panel que será construido.}

\item{capitulo}{una cadena de texto o un vector de cadenas de texto. Selecciona el o los capítulos en el marco de datos.}
}
\value{
un marco de datos tibble.
}
\description{
Construye paneles renombrando las variables para el acople entre periodos.
}
\examples{
edit_panel(panel = "pii", capitulo = "v")
edit_panel(panel = "pi", capitulo = "todos")

}
