% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_renombre.R
\name{edit_renombre}
\alias{edit_renombre}
\title{Plantilla para renombre de variables}
\usage{
edit_renombre(panel = NULL, capitulo = NULL)
}
\arguments{
\item{panel}{una cadena de texto. Selecciona la plantilla de renombre para el panel especificado.}

\item{capitulo}{una cadena de texto o un vector de cadenas de texto. Selecciona la plantilla de renombre para el capítulo o los capítulos del panel especificado.}
}
\value{
un marco de datos tibble.
}
\description{
Plantilla para renombre de variables
}
