% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_capitulo.R
\name{edit_capitulo}
\alias{edit_capitulo}
\title{Datos EDIT-S según capítulo y periodo}
\usage{
edit_capitulo(periodo, capitulo)
}
\arguments{
\item{periodo}{una cadena de texto. Selecciona el periodo en el marco de datos. Este argumento puede tomar el valor `"ii"`, `"iii"`, `"iv"` o `"v"`.}

\item{capitulo}{una cadena de texto o un vector de cadenas de texto. Selecciona el o los capítulos en el marco de datos. Puede tomar los valores `"i"`, `"ii"`, `"iii"`, `"iv"`, `"v"` o `"vi"`.}
}
\value{
un marco de datos tibble.
}
\description{
Invoca los datos de la Encuesta de Desarrollo e Innovación Tecnológica en Servicios (EDIT-S) seleccionando los capítulos y períodos indicados.
}
\examples{
edit_capitulo(capitulo = "ii", periodo = "iii")
edit_capitulo(capitulo = c("i","iv"), periodo = "v")

}
